[bug-fixing-failing-tests] updating failing tests
